Moving all classes to Builder folder but internals
