+++ src/JWT/Builder/Internal/EnumHelper.cs
